Fixed and improved aggregation code. Removed the Excel output. Added datatype conversion into univ3_pool.py
Cleaned up aggregate_5min.py by using constants from config.py. Also added comments for readability. Took care of unnecessary conversions. Removed Excel output, as I do not think it is worth it, as it was mainly there for myself to check through stuff easier. It is still in commit history, so I would refer to that. Might add it later. Added float conversion for a query in univ3_pool.py for better readability in the aggregate_5min.py

+++ code/data/univ3_pool.py
@@ -78,15 +78,15 @@ def query_pool(window_end_str):
     tick  = int(pool["tick"]) if pool["tick"] is not None else None
     row   = {
         "window_end":    window_end_str,
-        "liquidity":     pool["liquidity"],
+        "liquidity":     int(pool["liquidity"]),
         "tick":          tick,
         "sqrt_price":    pool["sqrtPrice"],
-        "price":         round(price, 8),
-        "token0_price":  pool["token0Price"],
-        "token1_price":  pool["token1Price"],
-        "tvl_token0":    pool["totalValueLockedToken0"],
-        "tvl_token1":    pool["totalValueLockedToken1"],
-        "tvl_usd":       pool["totalValueLockedUSD"],
+        "price":         float(round(price, 8)),
+        "token0_price":  float(pool["token0Price"]),
+        "token1_price":  float(pool["token1Price"]),
+        "tvl_token0":    float(pool["totalValueLockedToken0"]),
+        "tvl_token1":    float(pool["totalValueLockedToken1"]),
+        "tvl_usd":       float(pool["totalValueLockedUSD"]),
     }
     return row, tick
